2023: day 3: reset parsed number after each loop
This was causing the too high bug. Example input:

    425..324+

The code would parse 425, then land on the period character, find that
it was not a part, and then continue... without resetting the parsed
number. Once 324 was reached, the parsed num would be 425324, and since
it was recognized as a part, the code would add the giant number
(425324) instead of the correct part number (324). Oops!

Now my generated solution is somehow _lower_ than the answer :( But
still progress!